Create the foundation of the project
I created the HTML file and the CSS file. I then worked to structure
the HTML file as close as possible adding all the items/sections that
I thought were necessary to start the project. I am starting to move
into building the CSS file and format the landing page.